src/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.52.0
redis[hiredis]==5.2.1
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
sse-starlette==2.2.1
